// SCRTSCYT/convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";
import type * as auth from "../auth.js";
import type * as folderWatcher from "../folderWatcher.js";
import type * as folderWatcherMutations from "../folderWatcherMutations.js";
import type * as http from "../http.js";
import type * as movies from "../movies.js";
import type * as router from "../router.js";
import type * as seedData from "../seedData.js";
import type * as upload from "../upload.js";

/**
 * A utility for referencing Convex functions in your app's API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  folderWatcher: typeof folderWatcher;
  folderWatcherMutations: typeof folderWatcherMutations;
  http: typeof http;
  movies: typeof movies;
  router: typeof router;
  seedData: typeof seedData;
  upload: typeof upload;
}>;
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;